requirements.txt
torch>=2.0.0
numpy>=1.24.0
gymnasium>=0.28.0
matplotlib>=3.7.0
tensorboard>=2.12.0
tqdm>=4.65.0
pyyaml>=6.0
seaborn>=0.12.0
pandas>=2.0.0
